[FLINK-28569][table-planner] Move non-deterministic test functions to userDefinedScalarFunctions
